+++ ifit/pom.xml
@@ -2,11 +2,11 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>br.com.ifit.web</groupId>
-	<artifactId>ifit</artifactId>
+	<groupId>br.com.uaiGym.web</groupId>
+	<artifactId>uaiGym</artifactId>
 	<packaging>war</packaging>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>ifit Maven Webapp</name>
+	<name>uaiGym Webapp</name>
 	<url>http://maven.apache.org</url>
 	<dependencies>
 		<dependency>
@@ -37,6 +37,6 @@
 		<maven.compiler.target>1.8</maven.compiler.target>
 	</properties>
 	<build>
-		<finalName>ifit</finalName>
+		<finalName>uaiGym</finalName>
 	</build>
 </project>
