+++ ifit/src/main/java/uaiGym/controller/CadastroServlet.java
@@ -1,4 +1,4 @@
-package ifit.controller;
+package uaiGym.controller;
 
 import java.io.IOException;
 
@@ -9,10 +9,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import ifit.service.AuthService;
+import uaiGym.service.AuthService;
 
-@WebServlet("/")
-public class LoginServlet extends HttpServlet {
+@WebServlet("/cadastrar")
+public class CadastroServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
        
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -24,7 +24,7 @@ public class LoginServlet extends HttpServlet {
 			dispatcher = request.getRequestDispatcher("/WEB-INF/menu.jsp");
 		} else {
 			request.setAttribute("mensagem", authenticator.getMessages());
-			dispatcher = request.getRequestDispatcher("index.jsp");
+			dispatcher = request.getRequestDispatcher("/WEB-INF/cadastro.jsp");
 		}
 		
 		dispatcher.forward(request, response);
@@ -35,13 +35,19 @@ public class LoginServlet extends HttpServlet {
 		RequestDispatcher dispatcher;
 		AuthService authenticator = new AuthService(request.getSession());
 		
+		String email = request.getParameter("email");
 		String usuario = request.getParameter("usuario");
 		String senha = request.getParameter("senha");
+		String confirmarSenha = request.getParameter("confirmarSenha");
 		
-		authenticator.login(usuario, senha);
+		authenticator.register(email, usuario, senha, confirmarSenha);
 		request.setAttribute("mensagem", authenticator.getMessages());
-		
-		dispatcher = request.getRequestDispatcher("/");
+
+		if (authenticator.isValid()) {
+			dispatcher = request.getRequestDispatcher("/login");
+		} else {
+			dispatcher = request.getRequestDispatcher("/cadastrar");
+		}
 		dispatcher.forward(request, response);
 	}
