+++ ifit/src/main/java/uaiGym/controller/LoginServlet.java
@@ -1,4 +1,4 @@
-package ifit.controller;
+package uaiGym.controller;
 
 import java.io.IOException;
 
@@ -9,10 +9,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import ifit.service.AuthService;
+import uaiGym.service.AuthService;
 
-@WebServlet("/cadastrar")
-public class CadastroServlet extends HttpServlet {
+@WebServlet("/")
+public class LoginServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
        
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -24,7 +24,7 @@ public class CadastroServlet extends HttpServlet {
 			dispatcher = request.getRequestDispatcher("/WEB-INF/menu.jsp");
 		} else {
 			request.setAttribute("mensagem", authenticator.getMessages());
-			dispatcher = request.getRequestDispatcher("/WEB-INF/cadastro.jsp");
+			dispatcher = request.getRequestDispatcher("index.jsp");
 		}
 		
 		dispatcher.forward(request, response);
@@ -35,19 +35,13 @@ public class CadastroServlet extends HttpServlet {
 		RequestDispatcher dispatcher;
 		AuthService authenticator = new AuthService(request.getSession());
 		
-		String email = request.getParameter("email");
 		String usuario = request.getParameter("usuario");
 		String senha = request.getParameter("senha");
-		String confirmarSenha = request.getParameter("confirmarSenha");
 		
-		authenticator.register(email, usuario, senha, confirmarSenha);
+		authenticator.login(usuario, senha);
 		request.setAttribute("mensagem", authenticator.getMessages());
-
-		if (authenticator.isValid()) {
-			dispatcher = request.getRequestDispatcher("/login");
-		} else {
-			dispatcher = request.getRequestDispatcher("/cadastrar");
-		}
+		
+		dispatcher = request.getRequestDispatcher("/");
 		dispatcher.forward(request, response);
 	}
