+++ ifit/src/main/java/uaiGym/controller/LogoutServlet.java
@@ -1,4 +1,4 @@
-package ifit.controller;
+package uaiGym.controller;
 
 import java.io.IOException;
 
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import ifit.service.AuthService;
+import uaiGym.service.AuthService;
 
 @WebServlet("logout")
 public class LogoutServlet extends HttpServlet {
