+++ ifit/src/main/java/uaiGym/model/enuns/EstadoEnum.java
@@ -1,4 +1,4 @@
-package ifit.model.enuns;
+package uaiGym.model.enuns;
 
 public enum EstadoEnum {
 	AC, AL, AP, AM, BA, CE, DF, ES, GO, MA, MT, MS, MG, PA, PB, PR, PE, PI, RJ, RN, RS, RO, RR, SC, SP, SE, TO;
