+++ ifit/src/main/java/uaiGym/model/enuns/SexoEnum.java
@@ -1,4 +1,4 @@
-package ifit.model.enuns;
+package uaiGym.model.enuns;
 
 public enum SexoEnum {
 	MASCULINO, FEMININO;
