+++ ifit/src/main/java/uaiGym/model/enuns/TipoContratoEnum.java
@@ -1,4 +1,4 @@
-package ifit.model.enuns;
+package uaiGym.model.enuns;
 
 public enum TipoContratoEnum {
 	ANUAL, SEMESTRAL, MENSAL, TRIAL;
